removed bug that caused flickering
my app only draws changes to the canvas. when the window resizes, it needs you to redraw the whole frame, otherwise the screen goes black, or if you present the buffer again it flickers. i wrote it so that now even if it werent for the PresentAsync on line 98 of DebugTerminalClient the currently available terminal buffer is fully redrawn to the screen. when the window gets bigger, the new space will remain empty until the buffer view is presented again